Inform when automatic cert generation has been completed
Otherwise there's no indication and it has, and the user may think it's
hung.

Using qInfo would be more semantically correct for this than qWarning,
but if we do that, there's a chance the user will never see it if they
have their logging levels set to show warning messages but not info
messages. So to make sure they do, let's use the same logging level as
the original message.

Fixes #1