PortalSession: "refcount" enablement state
We should disable the portal session when nothing is using it anymore,
that way the compositor won't be doing needless work. Since the portal
session is shared, we should only do this when there's no users anymore.
So keep a list of users and remove those if they disappear or ask for
disable and only disable if there are no more users.