Improve rendering latency and stability
We cannot drop random frames after the encoder. It leads to content
glitches or a frozen video stream. It's best to always send everything
and rely on the pre-encoder frame queue.

If we have multiple frames ready to send, send them immediately.
Otherwise when we change the framerate we don't burn through our queue
and our latency will never decrease.

This patch also takes on a more aggressive frame rate policy. We
currently
calculate the average throughput and send the target based on that. This
means
we never get through any existing backlog or provide a safety net.