kcm: use proper signal handlers for checkboxes and slider
We shouldn't use on[Thing]Changed() because it can fire for
non-user-initiated modifications. Instead use onToggled() for the
checkboxes and onMoved for the slider.

For the text fields, it's okay to use onTextChanged() since that
provides a better UX than onEdited(), as onEdited() requires the user
to hit Return/enter or change focus, and that's not obvious.