kcm: use ScrollViewKCM as the base component
This KCM has a header with some messages and a label in it, a scrollview
in the center, and a FormLayout with settings below that.

This sounds a suspiciously like what ScrollViewKCM was designed to
facilitate! The Virtual Desktops KCM already paves the way. So let's do
the same thing here.