kcm: fix existing username detection
Doing this after the Save button was clicked can't work, because by the
time that happens, the accepted() signal has gotten fired which
automatically closes the dialog.

Instead, check for existing usernames as the user types, and disable the
button when needed. Also, use bindings for this, rather than so much
declarative property assignments.