Adjust encoding stream buffer size to be time based
The current mechanism for handling timing in krdp is two-fold.

- VideoStream will adjust the target ratio of frames out of kwin and
entering our pipeline
- If the pipeline is too full the encoder will throw away pending
frames, so that we don't build up a backlog.

This second step should be based on the frame rate so that we queue up N
seconds of data, rather than N frames. If we are rendering 1fps, we
don't want a minutes worth of data in the queue. This keeps latency
consistent when the frame rate changes.