server: Create a single temporary portal session on server startup
We want to ask the user for permission when it's convenient, not when
there's nobody at the machine to confirm the permissions. So create a
portal session at startup to request the permisisons.

Unfortunately we cannot reuse this portal session since it leads to
reconnection issues with the PipeWire stream. While that is probably
solvable it also causes additional problems, most notably that the
"Remote Desktop" SNI will be constantly active. So instead, discard the
session after it was successfully started and create a new session when
a new connection is incoming.

This assumes the user checked "remember" for the portal permission.
Unfortunately there is not a whole lot we can do to force that. In
general the permission model here is rather lacking.

CCBUG: 488372