server: Properly cleanup SessionWrapper objects
Since we're storing the connection as a QPointer comparing the destroyed
connection to the stored connection doesn't work since the QPointer is
already cleared at the point the connection has been destroyed. We don't
really want to track the connection anyway, we want to track the
wrapper. So make the wrapper emit a signal on connection loss that we
use to cleanup the wrapper.

BUG: 488367
BUG: 488364
(cherry picked from commit 090d741331b84462530b9d307d31a482cfd68973)