SessionController: remove SNI status; keep it always passive
Since the portal SNI is the one indicating if connection is active
or not, remove the SNI status. No need for two SNIs to be active.

Ideally we only would have one anyway, but this a workaround
until we can reuse the SNI for the portal status as well.

BUG: 488365
FIXED-IN: 6.1.1


(cherry picked from commit a4f25f535035601cb11cf7d494d15e5c675a0d10)

75114bbd SessionController: remove sni status, keep it always passive

Co-authored-by: Akseli Lahtinen <akselmo@akselmo.dev>