krdpserver: Add dummy clipboard to avoid crashing
This adds an unimplemented dummy clipboard
object to the krdpserver so that it won't
crash when cliprdr channel is enabled by the client.

It currently does nothing but starts and stops,
it does not support any copying or pasting yet.

This needs to be implemented on later date.


(cherry picked from commit 6f2933bb1dce459be0c17575ca180fd24b352693)

297dd9b3 krdpserver: Add dummy clipboard to avoid crashing
68f50d22 Move clipboard channel test inside drdynvc state check

Co-authored-by: Akseli Lahtinen <akselmo@akselmo.dev>