kcm: Set new quality default to 75, minimum quality to 50
The qualities below 50 are rather poor and any gained responsiviness
does not really help when you can't see what you're doing.

Instead, set the minimum quality value to 50, where it's still legible.
Set default quality to 75, which is around middle of the slider, and
should work well for most cases.