Move initial auth token fetching to KCM
On startup the krdp-server creates a stream request, this is because the
first prompt we show a user-facing dialog. It is much better for this to
happen whilst the user is setting up the remote desktop server, rather
than when remotely accessing the first time and not near the computer.

This feature got broken at some point, now it's fixed it's problematic
on next upgrade users might get a notification that recording is
happening with absolutely no interaction. This is confusing and bad.

All of this is a temporary measure with a better solution being put into
our xdg-desktop-portal-kde (2d4a4dfec07e478ab57fb70a1f531ba496382467)
but it is blocked on upstream.