Ship a Systemd preset to disable the service by default
Systemd leaves the default enablement state of user units up to the
distro. If a distro enables them all by default, that will include KRDP,
which is not desirable and will cause the actual enablement state to be
out of sync with the KCM.

Instead, ship our own .preset file to make sure the service is always
disabled by default on all Systemd-using distros.

BUG: 508477
FIXED-IN: 6.5.0