Improve the lacklustre system tray icon, make it more descriptive
If you aren't aware, KRdp creates a system tray icon which is somewhat
useful to tell if the server is currently running. However it has a few
problems I noticed which I then try to address in this patch:

1. The system tray icon is named "RDP Server" but the rest of the
user-facing language - the KCM specifically - uses the term "Remote
Desktop" which is much clearer.
2. There is no tooltip for the system tray icon, *I* know it's there
because the server is running but it doesn't actually say so.
3. There's no way to configure the server from the tray icon, the only
option is to stop it.